server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Coders for Causes"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
psycopg = {extras = ["binary", "pool"], version = "^3.2.1"}
freezegun = "^1.5.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
django-extensions = "^3.2.3"
django-filter = "^25.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
flake8 = "6.1.0"
flake8-django = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
